serde_yml

Module de

The de module contains the library’s YAML deserializer.

Structs

CurrentEnum

Represents the current state of an enum during deserialization.

Deserializer

A structure that deserializes YAML into Rust values.

DocumentAnchor

A structure that describes anchors and aliases in a YAML document. The anchor name, prefixed in the YAML document with “&”, is represented without the prefix in anchor_name. The anchor_path is a string that denotes the path to the anchor in the YAML document. Each key that form the path is separated from one another by “/”. The aliases vector contains the path to each reference to the anchor in the YAML document.

Enums

Event

Represents the different events that can occur during YAML parsing.

Progress

Represents the progress of parsing a YAML document.

Functions

from_reader

Deserialize an instance of type T from an IO stream of YAML.

from_slice

Deserialize an instance of type T from bytes of YAML text.

from_str

Deserialize an instance of type T from a string of YAML text.