serde_yml
Function from_value

pub fn from_value<T>(value: Value) -> Result<T, Error>
where
    T: DeserializeOwned,

Interpret a serde_yml::Value as an instance of type T.

This conversion can fail if the structure of the Value does not match the structure expected by T, for example if T is a struct type but the Value contains something other than a YAML map. It can also fail if the structure is correct but T’s implementation of Deserialize decides that something is wrong with the data, for example required struct fields are missing from the YAML map or some number is too big to fit in the expected primitive type.

let val = Value::String("foo".to_owned());
let s: String = serde_yml::from_value(val).unwrap();
assert_eq!("foo", s);