Module serde_yml::modules::error

The error module contains the Error struct and its implementation.

Structs

Error

An error that occurred during YAML serialization or deserialization.

Location

The input location where an error occurred.

Pos

Represents a position in the YAML input.

Enums

ErrorImpl

The internal representation of an error.

Functions

fix_mark

Fixes the mark and path in an error.

new

Creates a new Error from the given ErrorImpl.

shared

Creates a new Error from a shared ErrorImpl.

Type Aliases

Result

Alias for a Result with the error type serde_yml::Error.