serde_yml

Trait Index

pub trait Index: Sealed { }

A type that can be used to index into a serde_yml::Value. See the get and get_mut methods of Value.

This trait is sealed and cannot be implemented for types outside of serde_yml.