serde_yml
Module value

The value module contains the Value type for YAML values. The Value enum, a loosely typed way of representing any valid YAML value.

Re-exports

pub use self::tagged::Tag;

pub use self::tagged::TaggedValue;

pub use crate::number::Number;

Modules

tagged

A representation of YAML’s !Tag syntax, used for enums.

Structs

Mapping

A YAML mapping in which the keys and values are both serde_yml::Value.

Serializer

Serializer whose output is a Value.

Enums

Value

Represents any valid YAML value.

Traits

Index

A type that can be used to index into a serde_yml::Value. See the get and get_mut methods of Value.

Functions

from_value

Interpret a serde_yml::Value as an instance of type T.

to_value

Converts a serializable value into a serde_yml::Value.

Type Aliases

Sequence

A YAML sequence in which the elements are serde_yml::Value.