serde_yml::with::singleton_map_optional
Function serialize

pub fn serialize<T, S>(
    value: &Option<T>,
    serializer: S,
) -> Result<S::Ok, S::Error>
where
    T: Serialize,
    S: Serializer,

Serializes an optional value using the singleton_map representation.

If the value is Some, it will be serialized using the singleton_map representation. If the value is None, it will be serialized as null.

Arguments

value - A reference to the optional value to be serialized.

serializer - The serializer to use for serializing the value.

Returns

A result containing the serialization output or an error if serialization fails.