serde_yml
Module de

The de module contains the library’s YAML deserializer.

Structs

CurrentEnum

Represents the current state of an enum during deserialization.

Deserializer

A structure that deserializes YAML into Rust values.

Enums

Event

Represents the different events that can occur during YAML parsing.

Progress

Represents the progress of parsing a YAML document.

Functions

from_reader

Deserialize an instance of type T from an IO stream of YAML.

from_slice

Deserialize an instance of type T from bytes of YAML text.

from_str

Deserialize an instance of type T from a string of YAML text.