serde_yml::libyml

Module parser

YAML parser.

Structs

Anchor

Represents an anchor in a YAML document.

MappingStart

Represents the start of a mapping in a YAML document.

Parser

Represents a YAML parser.

ParserPinned

Represents a pinned parser for YAML deserialization.

Scalar

Represents a scalar value in a YAML document.

SequenceStart

Represents the start of a sequence in a YAML document.

Enums

Event

Represents a YAML event encountered during parsing.

ScalarStyle

Represents the style of a scalar value in a YAML document.