Function serde_yml::libyml::safe_cstr::debug_lossy

pub fn debug_lossy(bytes: &[u8], formatter: &mut Formatter<'_>) -> Result

Debugs a C string by printing it in a format that can be used in debugging output.

Arguments

bytes - A reference to the byte slice that represents the C string.

formatter - A mutable reference to the formatter where the debugged string will be written.

Returns

A Result indicating whether the debugging was successful.

Panics

This method will panic if the input bytes slice does not have a null terminator.