Module serde_yml::mapping

The mapping module contains the Mapping type for YAML mappings. A YAML mapping and its iterator types.

Structs

DuplicateKeyError

Error returned when a duplicate key is encountered while deserializing a serde_yml::Mapping.

IntoIter

Iterator over serde_yml::Mapping by value.

IntoKeys

Iterator of the keys of a serde_yml::Mapping.

IntoValues

Iterator of the values of a serde_yml::Mapping.

Iter

Iterator over &serde_yml::Mapping.

IterMut

Iterator over &mut serde_yml::Mapping.

Keys

Iterator of the keys of a &serde_yml::Mapping.

Mapping

A YAML mapping in which the keys and values are both serde_yml::Value.

OccupiedEntry

A view into an occupied entry in a Mapping. It is part of the Entry enum.

VacantEntry

A view into a vacant entry in a Mapping. It is part of the Entry enum.

Values

Iterator of the values of a &serde_yml::Mapping.

ValuesMut

Iterator of the values of a &mut serde_yml::Mapping.

Enums

Entry

Entry for an existing key-value pair or a vacant location to insert one.

Traits

Index

A trait for types that can be used to index into a serde_yml::Mapping.