serde_yml::mapping
Trait Index

pub trait Index: Sealed { }

A trait for types that can be used to index into a serde_yml::Mapping.

The get, get_mut, contains_key, remove, remove_entry, shift_remove and shift_remove_entry methods of Mapping use this trait to provide a uniform interface for indexing with different key types.

This trait is sealed and cannot be implemented for types outside of serde_yml.