serde_yml
Module with

The with module contains the With type for YAML values. Customizations to use with Serde’s #[serde(with = …)] attribute.

Modules

nested_singleton_map

Serialize/deserialize nested enums using a YAML map containing one entry in which the key identifies the variant name.

singleton_map

Serialize/deserialize an enum using a YAML map containing one entry in which the key identifies the variant name.

singleton_map_optional

Serialize/deserialize an optional enum using a YAML map containing one entry in which the key identifies the variant name.

singleton_map_recursive

Apply singleton_map to all enums contained within the data structure.

singleton_map_with

Serialize/deserialize an enum using a YAML map containing one entry in which the key identifies the variant name, while allowing combination with other serialize_with attributes.